refactor: rename toast-injector.ts -> toast-ref.ts
Since the Toastinjector was removed in the previous commit, the file name becomes inaccurate. This commit renames the toast-injector.ts to toast-ref.ts to better reflect the content of the file.
