+++ src/lib/public_api.ts
@@ -3,7 +3,7 @@ export * from './toastr/toast.component';
 export * from './toastr/toastr.service';
 export * from './toastr/toastr-config';
 export * from './toastr/toastr.module';
-export * from './toastr/toast-injector';
+export * from './toastr/toast-ref';
 export * from './toastr/toast-noanimation.component';
 
 export * from './portal/portal';
